refactor para poner el header en una unica seccion utils compartida por todos los views
